# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Skills4Attendance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cohorts:
    get:
      operationId: listCohorts
      tags: [cohorts]
      parameters:
        - name: tutorId
          in: query
          schema: { type: integer }
        - name: active
          in: query
          schema: { type: boolean }
        - name: programme
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of cohorts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Cohort" }
    post:
      operationId: createCohort
      tags: [cohorts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CohortInput"
      responses:
        "201":
          description: Cohort created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cohort"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cohorts/{id}:
    get:
      operationId: getCohort
      tags: [cohorts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cohort detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CohortDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCohort
      tags: [cohorts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CohortUpdate"
      responses:
        "200":
          description: Cohort updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cohort"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cohorts/{id}/learners:
    get:
      operationId: getCohortLearners
      tags: [cohorts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Learners in a cohort
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Learner" }

  # ---------- Allocation ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    LearnerStatus:
      type: string
      enum: [active, withdrawn, completed, paused]
...
    DeliveryDay:
      type: string
      enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
...
    Learner:
      type: object
      required:
        [
          id,
          learnerRef,
          uln,
          firstName,
          lastName,
          email,
          employer,
          programme,
          level,
          startDate,
          plannedEndDate,
          status,
          tutorId,
          cohortId,
          tutorName,
          cohortName,
          externalSystemId,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        learnerRef: { type: string }
        uln: { type: ["string", "null"] }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: ["string", "null"] }
        employer: { type: ["string", "null"] }
        programme: { type: string }
        level: { type: string }
        startDate: { type: string, format: date }
        plannedEndDate: { type: ["string", "null"], format: date }
        status: { $ref: "#/components/schemas/LearnerStatus" }
        tutorId: { type: ["integer", "null"] }
        cohortId: { type: ["integer", "null"] }
        tutorName: { type: ["string", "null"] }
        cohortName: { type: ["string", "null"] }
        externalSystemId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Cohort:
      type: object
      required:
        [
          id,
          name,
          programme,
          level,
          tutorId,
          tutorName,
          deliveryDay,
          sessionStartTime,
          sessionEndTime,
          startDate,
          endDate,
          active,
          externalSystemId,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        programme: { type: string }
        level: { type: string }
        tutorId: { type: ["integer", "null"] }
        tutorName: { type: ["string", "null"] }
        deliveryDay: { $ref: "#/components/schemas/DeliveryDay" }
        sessionStartTime: { type: string }
        sessionEndTime: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: ["string", "null"], format: date }
        active: { type: boolean }
        externalSystemId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CohortDetail:
      allOf:
        - $ref: "#/components/schemas/Cohort"
        - type: object
          required: [learnerCount]
          properties:
            learnerCount: { type: integer }
    CohortInput:
      type: object
      required:
        [
          name,
          programme,
          level,
          deliveryDay,
          sessionStartTime,
          sessionEndTime,
          startDate,
        ]
      properties:
        name: { type: string, minLength: 1 }
        programme: { type: string, minLength: 1 }
        level: { type: string, minLength: 1 }
        tutorId: { type: integer }
        deliveryDay: { $ref: "#/components/schemas/DeliveryDay" }
        sessionStartTime: { type: string, minLength: 1 }
        sessionEndTime: { type: string, minLength: 1 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        active: { type: boolean, default: true }
        externalSystemId: { type: string }
    CohortUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        programme: { type: string, minLength: 1 }
        level: { type: string, minLength: 1 }
        tutorId: { type: ["integer", "null"] }
        deliveryDay: { $ref: "#/components/schemas/DeliveryDay" }
        sessionStartTime: { type: string, minLength: 1 }
        sessionEndTime: { type: string, minLength: 1 }
        startDate: { type: string, format: date }
        endDate: { type: ["string", "null"], format: date }
        active: { type: boolean }
        externalSystemId: { type: ["string", "null"] }
...